+++ source/Modules/AzureDevOpsDsc.Common/Resources/Functions/Public/AzDoVariableGroup/Remove-AzDoVariableGroup.ps1
@@ -37,6 +37,12 @@ Function Remove-AzDoVariableGroup
         return
     }
 
+    $otherProjects = @($vg.variableGroupProjectReferences | Where-Object { $_.projectReference.name -and $_.projectReference.name -ne $ProjectName } | ForEach-Object { $_.projectReference.name })
+    if ($otherProjects.Count -gt 0)
+    {
+        Write-Warning "[Remove-AzDoVariableGroup] Variable group '$VariableGroupName' is also shared with project(s): $($otherProjects -join ', '). Removing it from the owning project '$ProjectName' deletes it for all of them."
+    }
+
     $params = @{
         ApiUri          = 'https://dev.azure.com/{0}/' -f (Get-AzDoOrganizationName)
         ProjectId       = $project.id